Restore permissions for manager-role
controller-gen seems to have moved to a new architecture in
v0.2.0-beta.1. In hind-sight we probably shouldn't have moved so early,
but now that we're here, we need to perform work arounds.

We have simply copy pasted the previous role.yaml into the file that
should have been generated for us from the annotations on the
controller. When controller-gen v0.2.0 is no longer in beta, we can try
generating again.

[#166374440]
Co-authored-by: Jonathan Tsang <jtsang@pivotal.io>