Update workflow for patching controller image
We have decided that we don't need to patch the image digest during
development. CI will carry on pinning the digest to avoid clashing
with images that have been locally pushed.
When we deliver a story, we have to trigger the job `update-controller-image-digest`
so that when somebody runs `make deploy`, the latest image that was
delivered will be deployed.

[#167010562]
Co-authored-by: Aitor Cedres <acedres@pivotal.io>