Adds finalizer for immediate cluster deletion
We need to prep the RabbitmqCluster if we want to delete it but not wait for queues to
be synced. Here we delete the StatefulSet so there are no restarts. Then
we shutdown RabbitMQ on each node so the pre-stop hook terminates
immediately. Then we remove the finalizer so all remaining
RabbitmqCluster resources are garbage collected