Simplify image templating in Operator chart
There are a lot of helper functions that we are not using and were
initially copied in an attempt to be compliant with a proposal for
"Standard" Helm charts. In our case, kustomize does not work well with
placeholder strings, so we decided to move away from this proposal and
do what works for our use case.

Co-authored-by: Michal Kuratczyk <mkuratczyk@pivotal.io>