RetryOnConflict conflict in Pod deletion prep
We don't want to block and continuously requeue if we can't delete the
finalizer or add Pod labels. We should only error in this flow if we
can't remove the finalizer. Otherwise we should retry a few times and
then remove the finalizer and accept that Pod deletion will have to wait
on the Pod pre-stop hook to complete