Add a note to reviewers to remind about squashing
In the spirit of keeping our commit history relevant and navigable, reviewers should also review the commit list presented in a PR. If some or all commits can be squashed, they should suggest so. If not all commits should be squashed, the reviewer should suggest to rebase the branch and re-push.