Use MetalLB latest version and use a local Config
Kind-on-c repo is going to remove MetalLB. We have to use our own
configuration.

This does not work in Linux and unlikely is going to work on Mac. It
seems to be broken in the master branch too. I'm not deleting everything
altoghether just to preserve this knowledge and perhaps fix it one day.