Use MetalLB latest version and use a local Config (#128)
* Use MetalLB latest version and use a local Config

Kind-on-c repo is going to remove MetalLB. We have to use our own
configuration.

The make target will succeed and deploy MetalLB correcly in Kind.
However, the external IPs provided will not be reachable. It seems to be
a configuration of the local host network. In Mac it is not possible to
route traffic from the host to the containers; this is documented as a
Docker limitation for Mac.