Add statefulSet override
- new crd properties: spec.override.statefulSet, which is
our own definition of appsv1.StatefulSet, which allows users to provide
a patch that will be applied on top of generated statefulset definition
as an override.
- new type EmbeddedObjectMeta, which is a partially metav1.ObjectMeta;
it is used in StatefulSet, PodTemplateSpec, and PersistentVolumeClaim.
- custome types are needed because of the pruning behavior of
`preserveUnknownFields: false`, which is going to be the default
behavior for crd v1.
- all fields in the custom StatefulSetSpec are optional; including
selector, serviceName...etc