Prune unnecessary operator RBAC permissions
We add RBAC permissions to the operator via generator annotations in
rabbitmqcluster_controller.go. The length of the feedback loop for
testing which are actually necessary (e.g. build an image, deploy to a
registry or kind, run all the tests against it) is long and haphazard.
Some roles are required by underlying structures (e.g. the listing and
watching functions in the Reflector) so it's not just a question of
checking what API calls we make in the Reconcile function.
Historically, this has resulted in us over provisioning.