Refactor to always set rabbitmq-env.conf in config map (#218)
Previously we were setting rabbitmq-env.conf only when it was
non-empty. This resulted in a bug when rabbitmq-env.conf was
set and then becomes empty. The changes in rabbitmq-env.conf
were not refelected. We now always set the rabbitmq-env.conf
regardless of it's value. This would create an empty rabbitmq-env.conf
file and set the value to an empty string when it's not configured
by the user.

Co-authored-by: Harsha Chaudhary <harshac@vmware.com>