Refactor to extract out plugins
Both config map and controller were accessing the same method to find
the list of plugins. Also, the config map contained RequiredPlugins
which was accessed in Controller. As a part of this commit,
rabbitmqPlugins are abstracted out to not expose the default or required
plugins and return a final list of user provided plugins and default
plugins.