Remove duplication and fix flaky integration test (#259)
The annotations integration test relied on the events
emitted by the controller. However, this is not a reliable way
to test if the annotations are updated. K8s aggregates similar events
which were produced in a very short period of time. This results
in loss of some events which the tests were expecting. See
https://github.com/kubernetes/kubernetes/pull/16798 for more
information on this behaviour.