Remove operator Helm chart
We implement one step installation of the operator in #228.

In addition, 'kubectl rabbitmq install-cluster-operator' can be used to
install the operator.

We don't want to support another installation method via helm because it
requires maintenance and suffers from CRD upgrade issues.

The rabbitmq chart is still kept, in particular for KSM integration.