Use a test constant instead of importing
Because importing an internal library into the tests guarantees that the
test will always pass; this weakens the test slightly, as we want this
test to become red if we change the resource name because it would be a
user facing change i.e. a change in the expectation of what our product
creates.