Update registry references to new name
We renamed our DockerHub registry and have to reflect the
same in the manifests.

It also simplifies the manifest structure by
removing some patches that were always applied. If a patch is always
applied, it should be part of the resource itself.

Images are not set using kustomize 'image' field, instead of using a
patch. Overlays still need to use a patch because they change the image
pull policy.