Hacky way to remove embedded descriptions only
Previously we used maxDescLen to remove all descriptions (to limit the
size of the CRD). This approach keeps descriptions of our properties
while removing them from the `override` part of the CRD which contains
large embedded core types. Size of the CRD is a bit larger than with no
descriptions but still significantly smaller than with all descriptions.

Closes #369