Mount configuration files as projected keys
RabbitMQ images may come with 'vendor' configuration in the default
locations. If we mount our configuration as a directory, vendor
configuration won't be present. The plugins files must moved to a
different location because this file must be writable by the rabbitmq
process. Previously, we were copying the file into an empty dir volume
and passing this volume to the rabbitmq container. Since we moved away
from that behaviour, the plugins file has to be copied to a different
location.