Remove redundent and wrong assertion
- the expect assertion on length of imagePullSecrets is in
an eventually but not the returned result, which means that
the 3 seconds timeout does not apply and when failed at the
first iteration the assertion will just fail. While waiting
for the statefulSet to update, this could be very flaky on a
bad day
- the assertion itself is also redundent, as "ConsistOf" is
an exact matcher so the length is asserted already.