Rename additionalConfig to userDefinedConfiguration
The previous name could set an expectation to be a 1-to-1 mapping
between the file and the contents in `.spec.rabbitmq.additionalConfig`.
This was not intended, as this file can contain other configuration,
generated based on the user input.

Reverted the change that used string concatenation instead of multi-line
string in test expectations.