Place RabbitMQ configuration in confd (#548)
* Place RabbitMQ configuration in confd

Files places in confd are loaded after 'rabbitmq.conf' is loaded,
allowing the user configuration to take precedence. The ConfigMap has
keys without any numeric prefix because file naming is done in
StatefulSet's pod template volume mounts. This allows flexibility in the
future, shall we decide we rename or change the numeric prefixes.

This commit also adds a condition to decide whether to include the
server ConfigMap as a volume. This happens if RabbitMQ env or advanced
config are set in the RabbitmqCluster Spec. The server conf is partially
projected into a volume that is always specified.

* Use a string builder to generate RMQ configuration

String builder is more efficient than a byte buffer for strings.

* Rename additionalConfig to userDefinedConfiguration

The previous name could set an expectation to be a 1-to-1 mapping
between the file and the contents in `.spec.rabbitmq.additionalConfig`.
This was not intended, as this file can contain other configuration,
generated based on the user input.

Reverted the change that used string concatenation instead of multi-line
string in test expectations.