Expose operator metrics port
The operator metrics default port is 9782.
Enable operator metrics by default (which was already the case before
this commit.)

As of now, the operator metrics do not contain sensitive information.
Therefore, there is no need to use basic auth and a proxy to scrape the
metrics. Hence, remove the auth proxy patch file.

Include the metrics port in the mangager deployment YAML so that we can
create PodMonitors to point to the metrics port.
Add examples how to scrape the operator metrics.