Deploy to "examples" namespace
This example is namespace-sensitive. The `dnsNames` property needs to
contain the actual namespace this exmaple will be deployed to.

By `namespace: examples` we make sure it won't just apply to any
namespace and hopefully we'll bring this to the user's attention.

It also fixes our pipelines failures where we deploy examples to
"examples" namespace.

Fixes #618