Use ServiceMonitor instead of PodMonitor
Originally, we used PodMonitors to make Prometheus discover RabbitMQ
nodes to scrape.

Before this commit, when a user deloys both files rabbitmq-podmonitor.yaml
and rabbitmq-podmonitor-tls.yaml, Prometheus ends up scraping two
different instances (<host>:<port>) for the same RabbitMQ node for RabbitMQ
clusters where TLS is enabled but where disableNonTLSListeners is set to false.
In Grafana, this doubles the number of RabbitMQ nodes showing up.

As a solution, we could create a separate PodMonitor per RabbitMQ
cluster either including 15691 or 15692. This comes with the disadvanage
of having many PodMonitors and the user would need to change the
endpoint in the PodMonitor when TLS gets en- or disabled (since the
RabbitMQ cluster operator doesn't deploy the PodMonitors itself).

Here, we choose a different solution: We use a ServiceMontitor instead
of PodMonitors. A RabbitMQ service will either expose 15691 or 15692,
but never both ports. 15692 is exposed if disableNonTLSListeners is set
to true.

In this commit, we also delete the example for the kube-prometheus-stack
Helm chart since it was a duplicate.
We delete the separate TLS PodMonitor as well since we can now combine
both 15691 and 15692 ports in a single ServiceMonitor.

The ability to use ServiceMonitors have also been asked for by the
community, see:
* https://rabbitmq.slack.com/archives/CTMSV81HA/p1617990998094000
* https://rabbitmq.slack.com/archives/CTMSV81HA/p1611187267023900