Wider randomized_startup_delay_range
We've observed cluster formation failures at a rate of about 7%.
With these values, it seems to be below 1% (not a single failure in 100
tests). The ultimate solution remains to implement locking in the k8s
peer discovery plugin but this change should alleviate the issue until
then.

More info: https://github.com/rabbitmq/cluster-operator/issues/662