Allow override feature to behave more like a strategic merge patch
Previously, fields provided to the override feature were unmarshalled
early, meaning the patch that was applied was subject to the json
unmarshalling rules for the field types. Where those fields were core
API types, the operator does not have control over how those are
unmarshalled.

In this case, items marked `omitempty` could never be explicitly set to
`null`; for example, one could not set the securityContext of
RabbitmqClusters to their defaulted value.

This change allows for override patches to the core API types to behave
more like strategic merge patches:
- corev1.ServiceSpec
- corev1.PodSpec
- corev1.PersistentVolumeClaimSpec

This also fixes the bug where override patches had to specify
`containers: []` in the Pod Spec override despite needing no changes to
the containers.

This change is back-compatible.