Make PVC scaling more idempotent & resilient to crashes (#838)
* Extract PVC scaling to its own package

This allows for this section of the code to have its own test suite,
without relying on the restrictions placed by envtest. This allows us to
test more scenarios, such as where a controller has previously crashed
and left the existing API objects in a bad state (e.g. statefulsets
missing, PVCs being the wrong size, etc.).

This also fixes #782, as the process of scaling PVCs should now be more
idempotent than before:
- StatefulSets don't need to be present on the cluster in order to scale
the PVCs (as long as the PVCs exist)
- PVC scaling is done on a PVC-by-PVC case, rather than all at once, so
that if one or more replicas have already scaled their PVCs it does not
block the others
- The current size of PVCs is determined by the live reading of the K8s
API, rather than what's reported in the RabbitmqCluster object.

* Fix integration tests

* Move scaling test BeforeEach out of the test suite

* Fix nil pointer exception

when not defining replicas in Go struct

Before this commit, when replicas was not defined as in:
    Spec: rabbitmqv1beta1.RabbitmqClusterSpec{
            Replicas: &one,
    },

unit tests were failing.

Replicas is an optional field.
If not set, it should default to 1 instead of causing a panic.

* Do not pollute info logs

Before this commit huge "Found exising PVCs" logs
were output after every reconcilication even though PVCs didn't get
resized.

"By default logr's V(0) is zap's InfoLevel and V(1) is zap's DebugLevel (which is numerically -1)."

Co-authored-by: David Ansari <david.ansari@gmx.de>