Move default image from CRD spec to command line flag
Instead of the default image for RabbitmqClusters being hard-coded into
the CRD, it is now available as a command-line flag to the operator's
`manager` binary. Changing the operator's Deployment manifest to call
`/manager --default-rabbitmq-image image:tag` will now cause all
RabbitmqClusters deployed by the operator to use that default image if
one is not specified in `spec.image`.

This closes #856