Move default image from CRD spec to command line flag (#858)
* Move default image from CRD spec to command line flag

Instead of the default image for RabbitmqClusters being hard-coded into
the CRD, it is now available as a command-line flag to the operator's
`manager` binary. Changing the operator's Deployment manifest to call
`/manager --default-rabbitmq-image image:tag` will now cause all
RabbitmqClusters deployed by the operator to use that default image if
one is not specified in `spec.image`.

This closes #856

* Review markups

* Allow command-line parameters to just-run

This is useful to test specific values for the operator. For example, to
test different values to `-default-rabbitmq-image`.

Signed-off-by: Aitor Perez Cedres <acedres@vmware.com>

Co-authored-by: Aitor Perez Cedres <acedres@vmware.com>