Update kubectl-bats tests
The test to set debug logs was broken since the default image became
3.9.x because this commit: https://github.com/rabbitmq/rabbitmq-server/commit/cdcf602749a34674d9eb285df1e31fe9c4aebb89
changed the debug log prefix from `debug` to `dbug`.

There's a small tweak to the make target that runs the kubectl bats
tests locally to include the repo's bin first in the path. This is
important because the version installed by krew will take precedence.

The above also allows us to simplify slightly the PR workflow file and
run everything using `make`.

Signed-off-by: Aitor Perez Cedres <acedres@vmware.com>