Update Ginkgo to V2
Two significant changes affected our code base:

1. The table DSL i.e. `DescribeTable` and `Entry` have been moved to
   top-level. It was not required to import `extensions/table` anymore
1. Changes to CLI flag arguments to randomise all specs

There was a function name conflict because our code base and Ginkgo
define the function `Label()`. Due to the dot-import, it caused a name
conflict. This was resolved by aliasing our module.

Disabled metrics in integration tests to avoid MacOS firewall popup
everytime we run integration tests. We do not observe or scrape metrics
in the integration tests at the moment. In the future, we could bind to
localhost if needed.

Small refactor on getting the configuration to use a library function,
instead of a home-made one.

Signed-off-by: Aitor Perez Cedres <acedres@vmware.com>