Fix graceful shutdown
Previously all actions in preStop hook were performed independently of
each other. For example, drain would kick in even if queue
sinchronization returned an error. This resulted in cases where graceful
shutdown was assumed just because last action in the script, i.e.
`rabbitmq-upgrade drain`, is a success.

Now all actions in the hook actually waits for the previous action to
complete successfully. The only exception is checking for the deletion
marker file. Script will still exit with 0 (and shutdown will continue)
if the marker is present without executing other actions.