Fix examples
The examples test.sh script was not exiting when an error was returned.
Any error in the test.sh scripts should now stop the execution and
return the exit code.

__ Downscale vault-tls example
The github runners struggles to run 3-node rabbit cluster due to
resource constraints.

__ Delete obsolete example
Federation is better achieved using the Messaging Topology Operator.
There's another example that covers how to setup TLS. The case of
federation over TLS is a combination of using the Topology Operator in
combination with the TLS example.

__ Fix external secret example
The test.sh file was missing, and this test was not excluded from the CI
runs. Added a short script to validate that the external secret is used
to seed the default user.

__ Removed system tests in GKE
They do not provide more value than local KinD system tests

__ cmctl now waits watching the correct namespace
Earlier, if the kubeconfig had a namespace that didn't exist as current
ns, cmctl will always fail, even when cert-manager was ready

__ Skip mtls internode example tests

It requires a 3 node cluster, and this setup is not reliable inside a
GitHub runner, given the resource constraints inside the runner.

Typo in kubectl testss