[CI] Fix OLM CI
We stopped depending on a container image that was hosted in GCP. This
image is gone, and the workflow is now composed with small steps that do
something concrete. The tools required are either already available in
GitHub runners, or installed as part of a setup step. This ensures that
our tooling remains up to date.

The "publish" part now is separated in two different jobs, because their
steps are not dependant and they can be parallelised.

[skip ci]