[CI] OLM workflow updates
This commit changes the strategy to determine the OLM bundle version.
Previously, it was calculated in every job that needed this information.
Now, the bundle version is calculated in the first job that creates the
bundle, and passed on via outputs to subsequent dependant jobs.

The system tests were racy. Sometimes, they started too soon, before the
Cluster Operator was installed. Now, the system tests wait on the status
of Subscription. This ensures that Cluster Operator is, at least,
installed when the system tests execute.

[skip ci]