[CI] Remove replace version in OLM
The current logic is brittle, and it is guaranteed to break after we
release 2.10.0.

The problem with `ls -1v` is that it lists in alphabetical order, not
in semver order. That means that a list with 2.1.0, 2.9.0, 2.10.0 and
2.11.0 would be listed as:

2.1.0
2.10.0
2.11.0
2.9.0

The previous logic would always yield 2.9.0 (bottom) as the replace
version. That is not correct for 2.11.0.

The opm validator does not complain by not providing a "replaces"
version, so it may just be fine and OLM may have some magic sauce to
figure out what version to upgrade to.

[skip ci]