Remove ineffective test
This test became ineffective after #1735. We want to enable Web
MQTT/STOMP TLS regardless of whether a CA bundle certificate is
provided. This test was asserting that we enable web stomp/mqqt tls when
the CA bundle is set; however, by not setting the CA bundle, the test
passes anyway. This is not correct, making the test ineffective.