Update CodeQL workflow
Dependabot does not have write permissions in the repository, and it
causes failures when running on push events, because CodeQL needs write
access to upload the results of the scaning to GitHub.

This commit adds a conditional to run CodeQL if the event is a pull
request, OR if the actor is not dependabot. This conditional should
filter "push" events when dependabot is the actor, and always run pull
request scans.