Allow overriding readinessProbe and LivenessProbe
There can only be 1 probe and without this change,
specifying an exec readinessProbe in the spec.override would
add it on top of the default TCP probe.

As for livenessProbe, since there's currently no default,
everything works without this change. However, we could add
the default livenessProbe in the future and allowing this override
doesn't break anything right now, so it's just future-proofing.

Co-authored-by: joey <zchengjoey@gmail.com>