Move tools to dedicated module
This move allows to remove tool dependencies from the main mod file.
This is beneficial to reduce false positives in security scanners, and
to differentiate between CVEs in the operator code or its dependencies,
vs CVEs in the dependencies of the tools used to build the operator.

By moving the tools module to internal, we ensure that nobody can import
the tools module. We had seen a few users importing this project as a Go
library, even tho we never intended this project to be used as a library
:)