PROPOSAL: Reduce init container memory from 500Mi to 64Mi (#1863)
* Reduce init container memory from 500Mi to 64Mi

When originally developed, it wasn't clear what would need to be done in
the init container. With hindsight, there's very little to do and the
init phase hasn't changed in a long time.

We also assumed that the resources requested for the init container
would effectively be reused by the actual `rabbitmq` container later.
Turns out that's not the case: https://github.com/kubernetes/kubernetes/issues/124282
init container resources are considered even after the init container
completes its task.

* Reduce CPU resources from 100m to 20m