Add promtool as tool dependency in root go.mod
There is some incompatibility between our other tools and promtool. If
promtool is included as a tool in tools module, it will fail to build
with a compilation error, calling with the wrong number of arguments a
Kubernetes API function.

As workaround, promtool must be tracked in the root go.mod file.