Add `autoEnableAllFeatureFlags` property
Default to false. When set to true, all feature flags are automatically
enabled immediately after each upgrade (we don't check
if there are new feature flags - in most cases, this command
is a no-op). Note that when upgrade to RabbitMQ 4.2, this means
Khepri is gets enabled just after the upgrade
(again, if the property is set to true).