olm: refactor CI workflow
The environment we were using to test the OLM bundle is long gone.
Instead, we are going to deploy a KinD environment, install OLM, and
test that our bundle installs. This test includes creating a
subscription, creating a rabbitmq object and verifying that images have
"rabbitmq" in the name. It is naive, but this workflow can be extended
to run operator system tests.