Make container tool configurable
Most of the time, we just use `docker` for all our container needs.
However, sometime you need to use a different tool like `podman` or
`containerd`; for example, to test on OpenShift Local. OpenShift Local
comes with podman and it forces you to use Podman if you want to build
containers locally.